requirements-lite.txt
youtube-transcript-api>=0.6.0
openai>=1.0.0
streamlit>=1.28.0
chromadb>=0.3.0
yt-dlp>=2024.03.10
python-dotenv>=1.0.0
requests>=2.31.0
